// src/shared/container/providers/mail-provider/dtos/isend-mail-dto.ts
import IParseMailTemplateDTO from '../../mail-template-provider/dtos/iparse-mail-template-dto';

interface IMailContact {
  name: string;
  email: string;
}

export default interface ISendMailDTO {
  to: IMailContact;
  from?: IMailContact;
  subject: string;
  templateData: IParseMailTemplateDTO;
}